+++ assetconnection/http/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/assetconnection/http/HttpAssetConnection.java
@@ -23,15 +23,29 @@ import de.fraunhofer.iosb.ilt.faaast.service.assetconnection.http.provider.HttpV
 import de.fraunhofer.iosb.ilt.faaast.service.assetconnection.http.provider.config.HttpOperationProviderConfig;
 import de.fraunhofer.iosb.ilt.faaast.service.assetconnection.http.provider.config.HttpSubscriptionProviderConfig;
 import de.fraunhofer.iosb.ilt.faaast.service.assetconnection.http.provider.config.HttpValueProviderConfig;
-import de.fraunhofer.iosb.ilt.faaast.service.assetconnection.http.security.SelfSignedCertificateHandler;
+import de.fraunhofer.iosb.ilt.faaast.service.certificate.util.KeyStoreHelper;
 import de.fraunhofer.iosb.ilt.faaast.service.config.CoreConfig;
 import de.fraunhofer.iosb.ilt.faaast.service.exception.ConfigurationInitializationException;
 import io.adminshell.aas.v3.model.Reference;
+import java.io.File;
 import java.io.IOException;
 import java.net.Authenticator;
 import java.net.PasswordAuthentication;
 import java.net.http.HttpClient;
 import java.security.GeneralSecurityException;
+import java.security.KeyStore;
+import java.security.SecureRandom;
+import java.security.cert.CertificateException;
+import java.security.cert.X509Certificate;
+import java.util.ArrayList;
+import java.util.Enumeration;
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Stream;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.TrustManager;
+import javax.net.ssl.TrustManagerFactory;
+import javax.net.ssl.X509TrustManager;
 import org.apache.commons.lang3.StringUtils;
 
 
@@ -39,26 +53,32 @@ import org.apache.commons.lang3.StringUtils;
  * Implementation of {@link de.fraunhofer.iosb.ilt.faaast.service.assetconnection.AssetConnection} for the HTTP
  * protocol.
  *
- * <p>Following asset connection operations are supported:
+ * <p>
+ * Following asset connection operations are supported:
  * <ul>
  * <li>setting values (via HTTP PUT)
  * <li>reading values (via HTTP GET)
  * </ul>
  *
- * <p>Following asset connection operations are not supported:
+ * <p>
+ * Following asset connection operations are not supported:
  * <ul>
  * <li>subscribing to values
  * <li>executing operations
  * </ul>
  *
- * <p>This implementation currently only supports submodel elements of type {@link io.adminshell.aas.v3.model.Property}
+ * <p>
+ * This implementation currently only supports submodel elements of type {@link io.adminshell.aas.v3.model.Property}
  * resp. {@link de.fraunhofer.iosb.ilt.faaast.service.model.value.PropertyValue}.
  *
- * <p>This class uses a single underlying HTTP connection.
+ * <p>
+ * This class uses a single underlying HTTP connection.
  */
 public class HttpAssetConnection extends
         AbstractAssetConnection<HttpAssetConnection, HttpAssetConnectionConfig, HttpValueProviderConfig, HttpValueProvider, HttpOperationProviderConfig, HttpOperationProvider, HttpSubscriptionProviderConfig, HttpSubscriptionProvider> {
 
+    private static final String PROTOCOL_HTTPS = "https";
+    private static final String PROTOCOL_SSL = "SSL";
     private HttpClient client;
 
     public HttpAssetConnection() {
@@ -98,8 +118,10 @@ public class HttpAssetConnection extends
     @Override
     protected void doConnect() throws AssetConnectionException {
         try {
-            HttpClient.Builder builder = HttpClient.newBuilder()
-                    .sslContext(new SelfSignedCertificateHandler().createCustomSSLContext(config));
+            HttpClient.Builder builder = HttpClient.newBuilder();
+            if (PROTOCOL_HTTPS.equalsIgnoreCase(config.getBaseUrl().getProtocol())) {
+                builder = builder.sslContext(createCustomSSLContext());
+            }
             if (StringUtils.isNotBlank(config.getUsername())) {
                 builder = builder.authenticator(new Authenticator() {
                     @Override
@@ -117,6 +139,59 @@ public class HttpAssetConnection extends
         catch (IOException | GeneralSecurityException e) {
             throw new AssetConnectionException("error establishing HTTP asset connection", e);
         }
+    }
+
+
+    private SSLContext createCustomSSLContext() throws IOException, GeneralSecurityException {
+        List<X509Certificate> trustedCertificates = loadTrustedCertificates(config);
+        SSLContext sslContext = SSLContext.getInstance(PROTOCOL_SSL);
+        TrustManagerFactory defaultFactory = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
+        defaultFactory.init((KeyStore) null);
+        X509TrustManager defaultTrustManager = (X509TrustManager) Stream.of(defaultFactory.getTrustManagers())
+                .filter(x -> X509TrustManager.class.isAssignableFrom(x.getClass()))
+                .findFirst()
+                .orElseThrow(() -> new IllegalArgumentException("unable to find default trust manager"));
+        sslContext.init(null, new TrustManager[] {
+                new X509TrustManager() {
+                    @Override
+                    public void checkClientTrusted(X509Certificate[] chain, String authType) throws CertificateException {
+                        defaultTrustManager.checkClientTrusted(chain, authType);
+                    }
+
+
+                    @Override
+                    public void checkServerTrusted(X509Certificate[] chain, String authType) throws CertificateException {
+                        if (chain.length == 1 && trustedCertificates.contains(chain[0])) {
+                            return;
+                        }
+                        defaultTrustManager.checkServerTrusted(chain, authType);
+                    }
+
+
+                    @Override
+                    public X509Certificate[] getAcceptedIssuers() {
+                        return defaultTrustManager.getAcceptedIssuers();
+                    }
+                }
+        }, new SecureRandom());
+        return sslContext;
+    }
+
+
+    private List<X509Certificate> loadTrustedCertificates(HttpAssetConnectionConfig config) throws IOException, GeneralSecurityException {
+        List<X509Certificate> result = new ArrayList<>();
+        if (Objects.nonNull(config.getKeyStorePath()) && !config.getKeyStorePath().isEmpty()) {
+            var keyStore = KeyStoreHelper.loadKeyStore(new File(config.getKeyStorePath()), config.getKeyStorePassword());
+            Enumeration<String> aliases = keyStore.aliases();
+            while (aliases.hasMoreElements()) {
+                var alias = aliases.nextElement();
+                var certificate = keyStore.getCertificate(alias);
+                if (Objects.nonNull(certificate) && X509Certificate.class.isAssignableFrom(certificate.getClass())) {
+                    result.add((X509Certificate) certificate);
+                }
+            }
+        }
+        return result;
     }